+++ src/data.ts
@@ -1,3 +1,6 @@
+import type { LogHeaders } from './headers';
+import type { ManagedPointer, Wasm, WasmObject } from './wasm';
+
 export type ParserEvent =
 	| { kind: ParserEventKind.Event; data: undefined }
 	| { kind: ParserEventKind.MainFrame; data: MainFrame }
@@ -41,3 +44,56 @@ export type Stats = {
 	/** The approximate percentage of the log data parsed so far as a number in the range `[0,1]`. */
 	progress: number;
 };
+
+export class DataParser implements WasmObject, IterableIterator<ParserEvent> {
+	readonly #wasm: Wasm;
+	readonly #ptr: ManagedPointer<DataParser>;
+	readonly #headers: LogHeaders;
+	#done = false;
+
+	/** @internal */
+	constructor(wasm: Wasm, ptr: ManagedPointer<DataParser>, headers: LogHeaders) {
+		this.#wasm = wasm;
+		this.#ptr = ptr;
+		this.#headers = headers;
+	}
+
+	free() {
+		this.#ptr.free();
+	}
+
+	get isAlive(): boolean {
+		return this.#ptr.isAlive;
+	}
+
+	get headers(): LogHeaders {
+		return this.#headers;
+	}
+
+	stats(): Readonly<Stats> {
+		return this.#wasm.dataStats(this.#ptr.ptr);
+	}
+
+	[Symbol.iterator]() {
+		return this;
+	}
+
+	get done(): boolean {
+		return this.#done;
+	}
+
+	next(): IteratorResult<Readonly<ParserEvent>> {
+		if (this.#done) {
+			return { done: true, value: undefined };
+		}
+
+		const value = this.#wasm.dataNext(this.#ptr.ptr);
+
+		if (value === undefined) {
+			this.#done = true;
+			return { done: true, value: undefined };
+		}
+
+		return { done: false, value };
+	}
+}