+++ src/entry/sync.ts
@@ -1,3 +1,0 @@
-export * from './common';
-export * from '../sync';
-export { Parser as default } from '../sync';