+++ src/wasm/index.ts
@@ -10,9 +10,9 @@ import { freezeMap, freezeSet, unreachable } from '../utils';
 import type { RawPointer } from './pointers';
 import type { WasmSlice } from './slice';
 import type { OptionalWasmStr, WasmStr } from './str';
-import type { ParserEvent, Stats } from '../data';
-import type { InternalFrameDef } from '../headers';
-import type { DataParser, LogFile, LogHeaders } from '../sync';
+import type { DataParser, ParserEvent, Stats } from '../data';
+import type { LogFile } from '../file';
+import type { InternalFrameDef, LogHeaders } from '../headers';
 
 export type { RawPointer, ManagedPointer } from './pointers';